Relocate LeJEPALoss to lightly.models.modules.lejepa
Move LeJEPALoss and lejepa_invariance_loss out of
lightly/loss/lejepa_loss.py and into
lightly/models/modules/lejepa.py to match the structure
specified in issue comment at https://github.com/lightly-ai/lightly/issues/1888#issuecomment-3996087102

The new location groups all method-specific LeJEPA components
(the encoder wrapper, the invariance formula, and the combined
objective) in one file, while SIGReg stays in lightly.loss
as a reusable isotropic-Gaussian regularizer usable on its own.

Update the loss __init__ to drop the LeJEPALoss re-export,
expose LeJEPALoss via lightly.models.modules,
move the corresponding test classes to tests/models/modules/test_lejepa.py,
and update lightly.loss.rst (LeJEPALoss is now documented
through the lightly.models.rst .lejepa automodule section).
