+++ docs/source/lightly.loss.rst
@@ -32,11 +32,6 @@ lightly.loss
 .. autoclass:: lightly.loss.lejepa_loss.SIGReg
    :members:
 
-.. autoclass:: lightly.loss.lejepa_loss.LeJEPALoss
-   :members:
-
-.. autofunction:: lightly.loss.lejepa_loss.lejepa_invariance_loss
-
 .. autoclass:: lightly.loss.koleo_loss.KoLeoLoss
    :members:
